Added Date Time Dependecy JDO and Fixed Getall ticket

+++ CSUR/src/main/java/com/cmpe275/project/CSUR/services/TicketingService.java
@@ -2,8 +2,13 @@ package com.cmpe275.project.CSUR.services;
 
 import com.cmpe275.project.CSUR.dao.TicketingRepository;
 import com.cmpe275.project.CSUR.model.TicketDetails;
+import org.joda.time.DateTime;
+import org.joda.time.Period;
+import org.joda.time.format.DateTimeFormat;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 
 @Service
@@ -18,12 +23,38 @@ public class TicketingService implements Ticketing {
         TicketDetails ticket = ticketingRepository.save(ticketDetails);
     }
 
+
+    public boolean validateTicket(long ticketId)
+    {
+
+//        TicketDetails ticket = ticketingRepository.findOne(ticketId);
+//        DateTime ticketTime = DateTime.parse(ticket.getDeptTime(),
+//                DateTimeFormat.forPattern("dd/MM/yyyy HH:mm:ss"));
+//
+//        DateTime currentTime = new DateTime();
+//        Period p = new Period(currentTime, ticketTime);
+//        int minutes = p.getMinutes();
+//        if( minutes > 60)
+//            return true;
+//        return false ;
+         return true;
+    }
+
     @Override
-    public void cancelTicket(long ticketId) {
+    public boolean cancelTicket(long ticketId) {
+        if(validateTicket(ticketId)) {
+            ticketingRepository.delete(ticketId);
+            return true;
+        }
+        return false ;
 
-        TicketDetails ticket = ticketingRepository.findOne(ticketId);
-
-        ticketingRepository.delete(ticketId);
 
     }
+
+    @Override
+    public List<TicketDetails> getTickets(String userId) {
+        System.out.println(ticketingRepository.findAllByUserId(userId));
+        return ticketingRepository.findAllByUserId(userId);
+    }
+
 }
